Add Safe deployment callback to safe-ethers-lib
