+++ packages/safe-ethers-lib/src/index.ts
@@ -1,7 +1,6 @@
-import { EthersTransactionResult } from './contracts/GnosisSafe/GnosisSafeContractEthers'
 import { CreateProxyProps } from './contracts/GnosisSafeProxyFactory/GnosisSafeProxyFactoryEthersContract'
 import EthersAdapter, { EthersAdapterConfig } from './EthersAdapter'
-import { EthersTransactionOptions } from './types'
+import { EthersTransactionOptions, EthersTransactionResult } from './types'
 
 export default EthersAdapter
 export { EthersAdapterConfig, EthersTransactionOptions, EthersTransactionResult, CreateProxyProps }
