+++ packages/safe-ethers-lib/src/types.ts
@@ -1,5 +1,13 @@
+import { ContractTransaction } from '@ethersproject/contracts'
+import { BaseTransactionResult } from '@gnosis.pm/safe-core-sdk-types'
+
 export interface EthersTransactionOptions {
   from?: string
   gasLimit?: number | string
   gasPrice?: number | string
+}
+
+export interface EthersTransactionResult extends BaseTransactionResult {
+  transactionResponse: ContractTransaction
+  options?: EthersTransactionOptions
 }
